Resolved Issues #2, 3, and 4.
Untested authentication servers were down. We need a team of more then
one attempt to build two structures at the same time. And have a person
log off while the builder is building and see if the other player can
completed the structure.
Added the ability for anyone on the team to complete a structure.
Added more food items to the Sentry heals.