Code Examples for Mine Script Pathing
These were old code that I used for pathing in a previous version.
Hoping to use these examples for the mine to make sure the mine is
connected to the hub with rails by using the old code that checked for
water. We might always want to recreate these buildings for fast travel
systems as well.